Develop

+++ src/main/java/pl/petrusiewicz/ReservationSystem/controller/OrganizationController.java
@@ -3,47 +3,45 @@ package pl.petrusiewicz.ReservationSystem.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.petrusiewicz.ReservationSystem.model.ConferenceRoom;
 import pl.petrusiewicz.ReservationSystem.model.Organization;
 import pl.petrusiewicz.ReservationSystem.service.OrganizationService;
 
 import javax.validation.Valid;
 
 @RestController
-@RequestMapping
+@RequestMapping("/organizations")
 public class OrganizationController {
 
     @Autowired
     OrganizationService service;
 
-    @GetMapping("/organizations")
+    @GetMapping
     public ResponseEntity getAll() {
-        return ResponseEntity.ok(service.getAll());
+        return ResponseEntity.ok(service.findAll());
     }
 
-    @GetMapping("organization/id/{id}")
+    @GetMapping("/{id}")
     public ResponseEntity get(@PathVariable int id) {
-        int size = service.getAll().size();
-        if (id > 0 && id <= size) {
-            return ResponseEntity.ok(service.get(id));
+        if (service.existById(id)) {
+            return ResponseEntity.ok(service.findById(id));
         } else {
             return ResponseEntity.badRequest().body("Organizacja o ID: " + id + " nie istnieje");
         }
     }
 
-    @GetMapping("/organization/{name}")
-    public ResponseEntity findByName(@PathVariable String name){
-        if (service.isExist(name)){
+    @GetMapping(params = "name")
+    public ResponseEntity findByName(@RequestParam String name){
+        if (service.existByName(name)) {
             return ResponseEntity.ok(service.findByName(name));
         } else {
             return ResponseEntity.badRequest().body("Nie ma organizacji o nazwie " + name);
         }
     }
 
-    @PostMapping("/organization")
+    @PostMapping
     public ResponseEntity add(@Valid @RequestBody Organization organization) {
         organization.setName(organization.getName().trim());
-        if (!service.isExist(organization.getName())) {
+        if (!service.existByName(organization.getName())) {
             service.add(organization);
             return ResponseEntity.status(201).build();
         } else {
@@ -51,10 +49,24 @@ public class OrganizationController {
         }
     }
 
-    @DeleteMapping("/organization/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity remove(@PathVariable int id){
-        service.remove(id);
-        return ResponseEntity.ok().build();
+        if(service.existById(id)){
+            service.remove(id);
+            return ResponseEntity.status(200).build();
+        } else {
+            return ResponseEntity.badRequest().body("Organizacja o ID: " + id + " nie istnieje");
+        }
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity update(@PathVariable int id, @Valid @RequestBody Organization organization){
+        if (service.existById(id)){
+            service.update(id, organization);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.badRequest().body("Organizacja o ID: " + id + " nie istnieje");
+        }
     }
